requirements.txt
beautifulsoup4
tqdm
requests
PyYAML
requirements-parser
pip-search
findimports
